Move web std::fs replacement to steven_std
