+++ std/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "steven_std"
+version = "0.0.1"
+authors = [ "iceiix <ice_ix@protonmail.ch>" ]
+edition = "2018"
+
+[dependencies]
+log = { version = "0.4.6", features = ["std"] }
